fix: 全项目安全+正确性+系统性修复（第一性原理对抗性审计驱动）
基于对抗性审计分三维度批量修复，后端 589 测试零回归（7 xfail 为 P3 待办哨兵）。

## 安全修复（11 CRITICAL + 23 HIGH）
- voice.correct→confirm 注入链：corrections 改强类型 + event_type 白名单
- POS/采购资金竞态：pay/refund/resume/return 加 with_for_update()
- market_admin 角色检查失效：改读 _token_role，移除 owner
- media 上传扩展名白名单 + 分块读防 OOM
- environment 三端点加认证 + days 范围约束
- staff PIN 登录接入 rate_limiter + hmac.compare_digest
- accounts/ai_actions body:dict → Pydantic schema
- voice.upload 加大小/类型校验
- 供应商付款补 require_permission
- CSV 公式注入净化 _sanitize_csv_cell
- auth refresh 权限提升修复 + 旧 token 吊销

## 正确性修复（15 CRITICAL + 36 HIGH）
- voice_parser_v2 中文数字百千万解析（"三百"→300 而非 3100）
- reports 月报 day_cogs * 0 → 真实 COGS
- dynamic_pricing markdown 循环反向 + _floor_price min→max
- batch.py spoiled 幽灵状态 → FIFO 白名单 + FEFO 排序
- voice/offline_sync FIFO 短缺校验防库存变负
- 采购退货幂等键 + 核销原 purchase 流水
- twin 多城市异常 + backup pg_dump URL 修复
- 日结锁定全覆盖 + held 付款拒绝 + 对账时区 CST→UTC
- 盘点损耗用成本替代售价 + sku_id + N+1 消除
- 节假日系数/节气/降水量逻辑修正
- 员工停用后 JWT 即时失效 + 幂等中间件仅缓存 2xx

## P2 系统性工程
- schema 金额字段 float→DecimalNum（PlainSerializer 输出 float，前端零感知）
- models CHECK 约束（9 个关键枚举）+ ondelete 策略（10 关系）
- created_at nullable→False（33 处）+ Mapped float→Decimal（7 字段）
- unit_service 全程 Decimal + cache sentinel 支持 falsy + config 友好报错
- RLS 接线（仅 PG 方言激活）+ RedisBackend async 化

## 对抗性测试
- test_anomalies_api.py：修复绿色假象 + 语义断言 + scan 端到端
- test_operations_adversarial.py：攻击面对抗性测试（8 绿 + 9 xfail 哨兵）
- test_market_admin_api.py：角色检查 + 跨租户校验（新增 12 测试）
- staff/auth：refresh 旧 token 失效 + 员工角色保持 + PIN 限流

测试基线：589 passed, 7 xfailed

+++ qiantan-brain/backend/app/models/environment.py
@@ -1,6 +1,7 @@
 """Environment record model."""
 
 from datetime import date, datetime
+from decimal import Decimal
 
 import sqlalchemy as sa
 from sqlalchemy.orm import Mapped, mapped_column
@@ -14,10 +15,10 @@ class EnvironmentRecord(Base):
     id: Mapped[int] = mapped_column(sa.Integer, primary_key=True, autoincrement=True)
     date: Mapped[date] = mapped_column(sa.Date, nullable=False)
     city: Mapped[str] = mapped_column(sa.String(50), default="上海")
-    temp_high: Mapped[float | None] = mapped_column(sa.Numeric(5, 1))
-    temp_low: Mapped[float | None] = mapped_column(sa.Numeric(5, 1))
+    temp_high: Mapped[Decimal | None] = mapped_column(sa.Numeric(5, 1))
+    temp_low: Mapped[Decimal | None] = mapped_column(sa.Numeric(5, 1))
     weather_type: Mapped[str | None] = mapped_column(sa.String(30))
-    rainfall_prob: Mapped[float | None] = mapped_column(sa.Numeric(5, 1))
+    rainfall_prob: Mapped[Decimal | None] = mapped_column(sa.Numeric(5, 1))
     is_holiday: Mapped[bool] = mapped_column(sa.Boolean, default=False)
     holiday_name: Mapped[str | None] = mapped_column(sa.String(50))
     day_of_week: Mapped[int | None] = mapped_column(sa.Integer)
