+++ qiantan-brain/backend/app/models/preference.py
@@ -1,6 +1,7 @@
 """Merchant preference model."""
 
 import uuid
+from decimal import Decimal
 
 import sqlalchemy as sa
 from sqlalchemy.orm import Mapped, mapped_column
@@ -18,5 +19,5 @@ class MerchantPreference(Base):
     risk_profile: Mapped[str] = mapped_column(sa.String(20), default="neutral")
     voice_dialect: Mapped[str] = mapped_column(sa.String(30), default="mandarin")
     favorite_products: Mapped[list[int] | None] = mapped_column(sa.JSON, default=list)
-    avg_order_size: Mapped[float | None] = mapped_column(sa.Numeric(10, 2))
+    avg_order_size: Mapped[Decimal | None] = mapped_column(sa.Numeric(10, 2))
     preference_data: Mapped[dict] = mapped_column(sa.JSON, default=dict)
